section implementation

+++ src/database/migrations/2020_02_24_011259_create_sections_table.php
@@ -15,7 +15,7 @@ class CreateSectionsTable extends Migration
     {
         Schema::create('sections', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->tinyInteger('parent_id')->nullable();
+            $table->integer('parent_id')->nullable();
             $table->string('slug', 255);
             $table->string('url', 255);
             $table->string('name', 255);
